Refactor automata package
